自动化现成框架/paas5/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.3
decorator==5.1.1
dictdiffer==0.9.0
dynaconf==3.1.9
execnet==1.9.0
filelock==3.7.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
jsonpath==0.82
packaging==21.3
paramiko==2.11.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.28.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
toml==0.10.2
toml2json==0.1.0
tomlconfig==1.2.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
